Revert "refactor: move registry.lua to manage"
This reverts commit f87dd7205da39ed81987f91c01b66b4c8df69291.
